scripts/python/xml2vhdl-uo/xml2vhdl/requirements.txt
alabaster==0.7.12
babel==2.6.0
breathe==4.11.1
certifi==2018.11.29
chardet==3.0.4
colorlog==4.0.2
docutils==0.14
idna==2.8
imagesize==1.1.0
jinja2==2.10
logging==0.4.9.6
lxml==4.3.0
markupsafe==1.1.0
numpy==1.16.0
packaging==19.0
pygments==2.3.1
pyparsing==2.3.1
pytz==2018.9
pyyaml==3.13
releases==1.6.1
requests==2.21.0
semantic-version==2.6.0
six==1.12.0
snowballstemmer==1.2.1
sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
typing==3.6.6; python_version < '3.5'
urllib3==1.24.1

